Add the SideBar component
